ClearerVoice-Studio/train/target_speaker_extraction/requirements.txt
librosa==0.10.2.post1
soundfile==0.12.1
opencv-python==4.10.0.84
mir_eval==0.7
numpy==1.24.3
yamlargparse==1.31.1
scipy==1.10.1
tqdm==4.66.5
pesq==0.0.4
pystoi==0.4.1
tensorboard==2.14.0
torch-complex==0.4.4
rotary-embedding-torch==0.8.3
python_speech_features==0.6
typeguard==4.3.0
huggingface-hub==0.26.2
gdown==5.2.0